Extract the 2 new editor properties into a plugin
