+++ src/editor/SlateEditor.jsx
@@ -15,7 +15,6 @@ import {
   hasRangeSelection,
   toggleInlineFormat,
   toggleMark,
-  MIMETypeName,
 } from 'volto-slate/utils'; // fixSelection,
 import EditorContext from './EditorContext';
 
@@ -57,10 +56,6 @@ class SlateEditor extends Component {
     const defaultExtensions = slate.extensions;
     const raw = withHistory(withReact(createEditor()));
 
-    // TODO: also look for MIME Types in the files case
-    raw.dataTransferFormatsOrder = ['text/html', 'files', 'text/plain'];
-    raw.dataTransferHandlers = {};
-
     const plugins = [...defaultExtensions, ...this.props.extensions];
     const editor = plugins.reduce((acc, apply) => apply(acc), raw);
 
@@ -71,40 +66,6 @@ class SlateEditor extends Component {
 
     editor.getSavedSelection = this.getSavedSelection;
     editor.setSavedSelection = this.setSavedSelection;
-
-    const { insertData } = editor;
-
-    // TODO: update and improve comments & docs related to
-    // `dataTransferFormatsOrder` and `dataTransferHandlers` features
-    editor.insertData = (data) => {
-      if (editor.beforeInsertData) {
-        editor.beforeInsertData(data);
-      }
-
-      for (let i = 0; i < editor.dataTransferFormatsOrder.length; ++i) {
-        const x = editor.dataTransferFormatsOrder[i];
-        if (x === 'files') {
-          const { files } = data;
-          if (files && files.length > 0) {
-            // or handled here
-            return editor.dataTransferHandlers['files'](files);
-          }
-          continue;
-        }
-        const satisfyingFormats = data.types.filter((y) =>
-          new MIMETypeName(x).matches(y),
-        );
-        for (let j = 0; j < satisfyingFormats.length; ++j) {
-          const y = satisfyingFormats[j];
-          if (editor.dataTransferHandlers[x](data.getData(y), y)) {
-            // handled here
-            return true;
-          }
-        }
-      }
-      // not handled until this point
-      return insertData(data);
-    };
 
     return editor;
   }
