+++ src/editor/extensions/index.js
@@ -2,3 +2,4 @@ export * from './insertBreak';
 export * from './insertData';
 export * from './isInline';
 export * from './withDeserializers';
+export * from './withOrderedDataTransfers';
