+++ src/utils/index.js
@@ -9,6 +9,5 @@ export * from './ops';
 export * from './random';
 export * from './selection';
 export * from './volto-blocks';
-export * from './mime-types';
 
 // export * from './todo';
